common: Relocate geth genesis, state parsers to common, blockchain respectively
