+++ packages/blockchain/src/index.ts
@@ -1,3 +1,4 @@
 export { Blockchain } from './blockchain'
 export { CasperConsensus, CliqueConsensus, Consensus, EthashConsensus } from './consensus'
 export { BlockchainInterface, BlockchainOptions } from './types'
+export * from './utils'
